Added functions append, first, last, size, and expand.
Append adds an element to the end of the list. It will also increase
list size by 50% if it is full.

First returns the first element in the list or -1 if the list is empty.

Last returns the last element in the list or -1 if it is empty.

Size returns the total amount of storage locations available in list.

Expand is simply a private function which expands the list size by 50%.
It is meant to reduce code duplicity.

Updated JavaDoc documentation.

Added JUnit test cases.